api/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
httpx
apscheduler